requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==22.0
pycodestyle==2.10.0
pyotp==2.7.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
djangorestframework-simplejwt==5.2.2
pillow==10.2.0
